+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Macros `bydi-was-called{-with}` now accept an optional
+  {second,third} argument to clear the mocks for that very function.
+
 ## [v0.5.1]
 
 ### Added